@@ -2,13 +2,20 @@
 
 namespace App\Entity;
 
-use App\Repository\TrickRepository;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
+use App\Services\Slugify;
+use DateTimeInterface;
 use Doctrine\ORM\Mapping as ORM;
+use Doctrine\Common\Collections\Collection;
+use Doctrine\Common\Collections\ArrayCollection;
+use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
+use Symfony\Component\Validator\Constraints as Assert;
+
 
 /**
- * @ORM\Entity(repositoryClass=TrickRepository::class)
+ * @ORM\Entity(repositoryClass="App\Repository\TrickRepository")
+ * @UniqueEntity(fields="name",  message="This name is already in use")
+ * @UniqueEntity(fields="slug",  message="This slug is already in use")
+ * @ORM\HasLifecycleCallbacks()
  */
 class Trick
 {
@@ -17,59 +24,77 @@ class Trick
      * @ORM\GeneratedValue()
      * @ORM\Column(type="integer")
      */
-    private $id;
+    private int $id;
 
     /**
-     * @ORM\Column(type="string", length=255, unique=true)
+     * @ORM\Column(name="name", type="string", length=255, unique=true)
+     * @Assert\Length(min="5", max="10",minMessage="Le nom doit faire au minimum 5 caractères",maxMessage="Le nom ne doit pas faire plus de 10 caractères")
+     * @Assert\NotBlank()
+     * @Assert\NotNull()
+     * @Assert\Regex(
+     *     pattern="/[a-zA-Z._\p{L}-]{1,20}/",
+     *     message="Not valid name: juste letter for name"
+     * )
      */
-    private $title;
-
-    /**
-     * @ORM\Column(type="string", length=255, unique=true)
-     */
-    private $slug;
+    private string $name;
 
     /**
      * @ORM\Column(type="text")
+     * @Assert\Length(min="5", max="400",minMessage="Le nom doit faire au minimum 5 caractères",maxMessage="Le nom ne doit pas faire plus de 400 caractères")
+     * @Assert\NotBlank()
+     * @Assert\NotNull()
+     * @Assert\Regex(
+     *     pattern="/[a-zA-Z0-9._\p{L}-]{1,20}/",
+     *     message="Not valid description"
+     * )
      */
-    private $description;
+    private string $description;
 
     /**
      * @ORM\Column(type="datetime")
      */
-    private $created;
+    private DateTimeInterface $createdAt;
 
     /**
-     * @ORM\Column(type="datetime")
+     * @ORM\Column(type="datetime", nullable=true)
      */
-    private $lastUpdate;
+    private ?DateTimeInterface $updatedAt;
 
     /**
-     * @ORM\ManyToOne(targetEntity=Category::class, inversedBy="tricks")
-     * @ORM\JoinColumn(nullable=false)
+     * @ORM\Column(name="slug", type="string", length=255, unique=true)
      */
-    private $category;
+    private string $slug;
 
     /**
-     * @ORM\OneToMany(targetEntity=Comment::class, mappedBy="trick", cascade={"remove"}, orphanRemoval=true)
+     * @ORM\OneToMany(targetEntity="App\Entity\Image", mappedBy="trick" ,cascade={"persist", "remove"})
      */
-    private $comments;
+    private ?Collection $images;
+
 
     /**
-     * @ORM\OneToMany(targetEntity=Image::class, mappedBy="trick",  cascade={"persist", "remove"}, orphanRemoval=true)
+     * @ORM\OneToMany(targetEntity="App\Entity\Video", mappedBy="trick", cascade={"persist", "remove"})
      */
-    private $images;
+    private ?Collection $videos;
+
 
     /**
-     * @ORM\OneToMany(targetEntity=Video::class, mappedBy="trick", cascade={"persist", "remove"}, orphanRemoval=true)
+     * @ORM\ManyToOne(targetEntity="App\Entity\Category", inversedBy="tricks")
+     * @Assert\NotNull()
+     * @Assert\NotBlank()
      */
-    private $videos;
+    private Category $category;
+
+    /**
+     * @ORM\OneToMany(targetEntity=Comment::class, mappedBy="trick")
+     */
+    private Collection $comments;
+
 
     public function __construct()
     {
-        $this->comments = new ArrayCollection();
-        $this->images = new ArrayCollection();
-        $this->videos = new ArrayCollection();
+       $this->videos = new ArrayCollection();
+       $this->images = new ArrayCollection();
+       $this->comments = new ArrayCollection();
     }
 
     public function getId(): ?int
@@ -77,14 +102,14 @@ class Trick
         return $this->id;
     }
 
-    public function getTitle(): ?string
+    public function getName(): ?string
     {
-        return $this->title;
+        return $this->name;
     }
 
-    public function setTitle(string $title): self
+    public function setName(string $name): self
     {
-        $this->title = $title;
+        $this->name = $name;
 
         return $this;
     }
@@ -101,36 +126,124 @@ class Trick
         return $this;
     }
 
-    public function getCreated(): ?\DateTimeInterface
+    public function getCreatedAt(): DateTimeInterface
     {
-        return $this->created;
+        return $this->createdAt;
     }
 
-    public function setCreated(\DateTimeInterface $created): self
+    public function setCreatedAt(DateTimeInterface $createdAt): self
     {
-        $this->created = $created;
+        $this->createdAt = $createdAt;
 
         return $this;
     }
 
-    public function getLastUpdate(): ?\DateTimeInterface
+    public function getUpdatedAt(): ?DateTimeInterface
     {
-        return $this->lastUpdate;
+        return $this->updatedAt;
     }
 
-    public function setLastUpdate(\DateTimeInterface $lastUpdate): self
+    public function setUpdatedAt(DateTimeInterface $updatedAt): self
     {
-        $this->lastUpdate = $lastUpdate;
+        $this->updatedAt = $updatedAt;
 
         return $this;
     }
 
-    public function getCategory(): ?Category
+
+    public function getSlug(): string
+    {
+        return $this->slug;
+    }
+
+
+    public function setSlug(string $slug): self
+    {
+        $this->slug = $slug;
+
+        return $this;
+    }
+
+    /**
+     * Initialize slug before persist or update
+     *
+     * @ORM\PrePersist
+     * @ORM\PreUpdate
+     */
+    public function initializeSlug(): void
+    {
+        if (empty($this->slug)) {
+            $slug = slugify::slug($this->name);
+            $this->setSlug($slug);
+        }
+    }
+
+    /**
+     * @return null|Collection
+     */
+    public function getImages(): ?Collection
+    {
+        return $this->images;
+    }
+
+    public function addImage(?Image $image): ?self
+    {
+        if (!$this->images->contains($image)) {
+            $this->images[] = $image;
+            $image->setTrick($this);
+        }
+
+        return $this;
+    }
+
+    public function removeImage(?Video $image): ?self
+    {
+        if ($this->images->contains($image)) {
+            $this->images->removeElement($image);
+            if ($image->getTrick() === $this) {
+                $image->setTrick(null);
+            }
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return null|Collection
+     */
+    public function getVideos(): ?Collection
+    {
+        return $this->videos;
+    }
+
+    public function addVideo(?Video $video): ?self
+    {
+        if (!$this->videos->contains($video)) {
+            $this->videos[] = $video;
+            $video->setTrick($this);
+        }
+
+        return $this;
+    }
+
+    public function removeVideo(?Video $video): ?self
+    {
+        if ($this->videos->contains($video)) {
+            $this->videos->removeElement($video);
+            if ($video->getTrick() === $this) {
+                $video->setTrick(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function getCategory(): Category
     {
         return $this->category;
     }
 
-    public function setCategory(?Category $category): self
+    public function setCategory(Category $category): self
     {
         $this->category = $category;
 
@@ -157,8 +270,7 @@ class Trick
 
     public function removeComment(Comment $comment): self
     {
-        if ($this->comments->contains($comment)) {
-            $this->comments->removeElement($comment);
+        if ($this->comments->removeElement($comment)) {
             // set the owning side to null (unless already changed)
             if ($comment->getTrick() === $this) {
                 $comment->setTrick(null);
@@ -168,78 +280,4 @@ class Trick
         return $this;
     }
 
-    /**
-     * @return Collection|Image[]
-     */
-    public function getImages(): Collection
-    {
-        return $this->images;
-    }
-
-    public function addImage(Image $image): self
-    {
-        if (!$this->images->contains($image)) {
-            $this->images[] = $image;
-            $image->setTrick($this);
-        }
-
-        return $this;
-    }
-
-    public function removeImage(Image $image): self
-    {
-        if ($this->images->contains($image)) {
-            $this->images->removeElement($image);
-            // set the owning side to null (unless already changed)
-            //if ($image->getTrick() === $this) {
-            //    $image->setTrick(null);
-            //}
-        }
-
-        return $this;
-    }
-
-    /**
-     * @return Collection|Video[]
-     */
-    public function getVideos(): Collection
-    {
-        return $this->videos;
-    }
-
-    public function addVideo(Video $video): self
-    {
-        if (!$this->videos->contains($video)) {
-            $this->videos[] = $video;
-            $video->setTrick($this);
-        }
-
-        return $this;
-    }
-
-    public function removeVideo(Video $video): self
-    {
-        if ($this->videos->contains($video)) {
-            $this->videos->removeElement($video);
-            // set the owning side to null (unless already changed)
-            //if ($video->getTrick() === $this) {
-            //    $video->setTrick(null);
-            //}
-        }
-
-        return $this;
-    }
-
-
-    public function getSlug(): ?string
-    {
-        return $this->slug;
-    }
-
-    public function setSlug(string $slug): self
-    {
-        $this->slug = $slug;
-
-        return $this;
-    }
 }
